online_features/requirements-htr.txt
Pillow==11.0.0
numpy==1.26.4
Faker==33.0.0
